return the current value immediately as a 3rd arg in useState.
In react hooks, this is the first arg. I'm concerned about binding
& changeability of this value over time, so I made useState return
a value-returning function instead, but this could be a very direct
contravention of how an iterating over time version of stateless-
functions might work: perhaps state ought be explicitly updated,
new calls of the function be done directly, with new functions
and handlers having to get set up & created when the state does
indeed change. That could be a big part of the point & picture.

Ergonomically if nothing else it would be nice to have the value
availabile.

I am tempted to switch to the getter being the 3rd arg, for react
hooks compatibility sake, not that I think this library will ever
really fit with react hooks code.