Pretty close on context. Out of scope provider still providing.
Ooh this is tricky. When a provider goes out of scope, I think
it is re-rendering. But the consumer is still maintaing it's original
view of the stack, even though the stack has changed. Recent work on
making stack immutable definitely causing issue: tempted to try to
fully remove but I think something in-between maybe needed?